// src/logic/serializers/enums/PlayerSerializerIds.ts
export enum PlayerSerializerIds {
  table = "players",
  tagColumn = "tag",
  userColumn = "userId",
  referredByColumn = "referredBy",
  referralLockColumn = "referralLock",
  lobbyCountColumn = "lobbyCount",
  lobbyCountUnrankedColumn = "lobbyCountUnranked",
  lobbyCountBotBashColumn = "lobbyCountBotBash",
  lobbyCount5v5Column = "lobbyCount5v5",
  lobbyCountReplayAnalysisColumn = "lobbyCountReplayAnalysis",
  offensesColumn = "offenses",
}